pyproject.toml
[project]
name = "cdr_bench"
version = "0.1.0"
description = "Scripts for performing dimensionality reduction on chemical datasets"
authors = [
    {name = "Alexey Orlov", email = "data2scale@gmail.com"}
]
dependencies = [
    "setuptools>=74.1.2",
    "h5py",
    "numpy",
    "pandas",
    "scipy>=1.14.1",
    "matplotlib>=3.9.2",
    "scikit-learn>=1.5.2",
    "numba>=0.60.0",
    "toml>=0.10.2",
    "seaborn>=0.13.2",
    "tqdm>=4.66.5",
    "rdkit>=2024.3.5",
    "openTSNE>=1.0.2",
    "chemographykit",
    "umap-learn>=0.5.6",
    "dgllife>=0.3.2",
    "lightning>=2.4.0",
    "pandarallel>=1.6.5",
    "ipywidgets>=8.1.5",
    "scikit-dimension>=0.3.4",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Documentation = "https://axelrolov.github.io/cdr_bench/"
Repository = "https://github.com/AxelRolov/cdr_bench"
"Bug Tracker" = "https://github.com/AxelRolov/cdr_bench/issues"
Paper = "https://onlinelibrary.wiley.com/doi/full/10.1002/minf.202400265"


[dependency-groups]
dev = [
    "jupyterlab>=4.2.5",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
]
test = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "pytest-xdist>=3.5",
]
lint = [
    "ruff>=0.9",
    "pre-commit>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "requires_rdkit: skip if rdkit unavailable",
    "requires_numba: skip if numba unavailable",
    "requires_gpu: skip if CUDA unavailable",
    "requires_tmap: skip if tmap unavailable",
    "requires_chemographykit: skip if chemographykit unavailable",
]
addopts = ["-ra", "--strict-markers", "--tb=short"]
filterwarnings = [
    "ignore::DeprecationWarning:numba",
    "ignore::FutureWarning:sklearn",
]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B905"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011", "SIM300"]
"**/__init__.py" = ["F401"]
"scripts/**" = ["F821", "F841", "B023", "B028", "W293", "RUF059", "SIM108", "SIM118", "SIM210"]
"src/cdr_bench/visualization/**" = ["E402", "F841", "RUF059", "SIM108", "SIM118", "UP038"]
"src/cdr_bench/features/chemdist_features.py" = ["F821"]
"src/cdr_bench/scoring/**" = ["B007", "SIM108", "SIM118", "RUF059"]
"src/cdr_bench/io_utils/**" = ["B028", "B904", "F841", "SIM102", "SIM108", "SIM118", "SIM210", "RUF059", "UP038"]
"src/cdr_bench/optimization/**" = ["F841", "RUF059", "SIM118"]
"src/cdr_bench/dr_methods/**" = ["B904", "SIM108"]

[tool.ruff.lint.isort]
known-first-party = ["cdr_bench"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
source = ["src/cdr_bench"]
omit = ["*/visualization/*", "*/chemdist_features.py"]

[tool.coverage.report]
show_missing = true
skip_covered = true
